progress

+++ components/consumers/defectdojo/client/client.go
@@ -64,7 +64,7 @@ func (client *Client) CreateFinding(
 	tags []string,
 	testID, line, cwe, foundBy int32,
 	falseP, duplicate, active bool,
-	cvssScore float64,
+	cvssScore float32
 ) (types.FindingCreateResponse, error) {
 	url := fmt.Sprintf("%s/findings", client.host)
 	body := types.FindingCreateRequest{
